feat: cache homepage data in Pinia store across navigations
